bloomerp/django_bloomerp/pyproject.toml
[project]
name = "django-bloomerp"
version = "0.1.0"
description = "Django bloomerp is "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "arabic-reshaper==3.0.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asgiref==3.8.1",
    "asn1crypto==1.5.1",
    "asttokens==3.0.0",
    "async-lru==2.0.4",
    "attrs==24.3.0",
    "babel==2.16.0",
    "beautifulsoup4==4.12.3",
    "binaryornot==0.4.4",
    "bleach==6.2.0",
    "Brotli==1.1.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "channels==4.1.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "comm==0.2.2",
    "cookiecutter==2.6.0",
    "crispy-bootstrap5==2024.2",
    "cryptography==43.0.1",
    "cssselect2==0.7.0",
    "daphne==4.1.2",
    "debugpy==1.8.9",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "Django==5.1.1",
    "django-browser-reload==1.18.0",
    "django-cotton==2.6.1",
    "django-crispy-forms==2.3",
    "django-extensions==3.2.3",
    "django-filter==24.3",
    "django-formtools==2.5.1",
    "django-htmx==1.19.0",
    "django-model-utils==5.0.0",
    "django-route-decorator==0.1.1",
    "django-tailwind-4==0.1.4",
    "djangorestframework==3.15.2",
    "docutils==0.21.2",
    "et-xmlfile==1.1.0",
    "executing==2.1.0",
    "fastjsonschema==2.21.1",
    "fonttools==4.54.0",
    "fqdn==1.5.1",
    "h11==0.14.0",
    "html5lib==1.1",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "idna==3.10",
    "importlib_metadata==8.5.0",
    "ipykernel==6.29.5",
    "ipython==8.30.0",
    "isoduration==20.11.0",
    "jaraco.classes==3.4.0",
    "jaraco.context==6.0.1",
    "jaraco.functools==4.1.0",
    "jedi==0.19.2",
    "Jinja2==3.1.4",
    "jiter==0.7.0",
    "json5==0.10.0",
    "jsonfield==3.1.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-events==0.10.0",
    "jupyter-lsp==2.2.5",
    "jupyter_client==8.6.3",
    "jupyter_core==5.7.2",
    "jupyter_server==2.14.2",
    "jupyter_server_terminals==0.5.3",
    "jupyterlab==4.3.3",
    "jupyterlab_pygments==0.3.0",
    "jupyterlab_server==2.27.3",
    "keyring==25.5.0",
    "langchain-core>=0.3.67,<1.0.0",
    "langchain-openai==0.2.12",
    "langgraph==0.2.60",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-sdk==0.1.47",
    "langsmith>=0.3.45,<1.0.0",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "more-itertools==10.5.0",
    "msgpack==1.1.0",
    "nbclient==0.10.1",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "nh3==0.2.18",
    "notebook==7.3.1",
    "notebook_shim==0.2.4",
    "numpy==2.1.1",
    "openai==1.57.2",
    "openpyxl==3.1.5",
    "orjson==3.10.12",
    "oscrypto==1.3.0",
    "overrides==7.7.0",
    "packaging==24.1",
    "pandas==2.2.2",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pillow==10.4.0",
    "pkginfo==1.10.0",
    "platformdirs==4.3.6",
    "plotly==5.24.1",
    "prometheus_client==0.21.1",
    "prompt_toolkit==3.0.48",
    "psutil==6.1.0",
    "psycopg2-binary==2.9.9",
    "pure_eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "pydyf==0.11.0",
    "Pygments==2.18.0",
    "pyHanko==0.25.1",
    "pyhanko-certvalidator==0.26.3",
    "pypdf==5.0.0",
    "PyPDF2==3.0.1",
    "pyphen==0.16.0",
    "pypng==0.20220715.0",
    "python-bidi==0.6.0",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.2.1",
    "python-slugify==8.0.4",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "pyzmq==26.2.0",
    "qrcode==7.4.2",
    "readme_renderer==44.0",
    "referencing==0.35.1",
    "regex==2024.11.6",
    "reportlab==4.2.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986==2.0.0",
    "rfc3986-validator==0.1.1",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "Send2Trash==1.8.3",
    "setuptools==75.0.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlparse==0.5.1",
    "stack-data==0.6.3",
    "svglib==1.5.1",
    "swapper==1.4.0",
    "tenacity==9.0.0",
    "terminado==0.18.1",
    "text-unidecode==1.3",
    "tiktoken==0.8.0",
    "tinycss2==1.3.0",
    "tornado==6.4.2",
    "tqdm==4.66.6",
    "traitlets==5.14.3",
    "twine==5.1.1",
    "types-python-dateutil==2.9.0.20241206",
    "typing_extensions==4.12.2",
    "tzdata==2024.1",
    "tzlocal==5.2",
    "uri-template==1.3.0",
    "uritools==4.0.3",
    "urllib3==2.2.3",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "xhtml2pdf==0.2.16",
    "zipp==3.21.0",
    "zopfli==0.2.3",
    "crispy-tailwind==1.0.3",
    "black>=26.1.0",
    "fa6-icons>=0.0.3",
    "celery>=5.4.0",
]
